fix sprites not being rendered on lowest possible floor
Prevent integer underflows by extending the integer value range.
Fixes #690, thanks @retro65.

+++ src/doom/r_things.c
@@ -713,7 +713,7 @@ void R_ProjectSprite (mobj_t* thing)
     // [JN] killough 4/9/98: clip things which are out of view due to height
     gzt = interpz + spritetopoffset[lump];
     if (interpz > viewz + FixedDiv(viewheight << FRACBITS, xscale) ||
-        gzt < viewz - FixedDiv((viewheight << FRACBITS)-viewheight, xscale))
+        gzt < (int64_t)viewz - FixedDiv((viewheight << FRACBITS)-viewheight, xscale))
     {
 	return;
     }
